Test without taste shocks for sim.
